update return grpc or json,docker, error redirect cors

+++ model/model.go
@@ -4,5 +4,5 @@ type URLEntry struct {
 	Id          int64  `json:"id"`
 	OriginalURL string `json:"original_url"`
 	ShortedURL  string `json:"short_url"`
-	Clicks      int64  `json:"clicks"`
+	Clicks      int32  `json:"clicks"`
 }
